Replaced "Make a Move" button with "Flip Board" button
The "Make a Move" button never really worked, and only double clicking
caused a move to happen, so I firstly removed it completely.  However,
this meant that the Resign button became the default button with the
unfortunate consequence that Resigning was almost automatic!  So, I've
added a "Flip Board" button, which currently does nothing, but is at least
less confusing than a "Make a Move" button that does nothing!