Added integrity checks to stop gamer writing -1 to db
There are some problems where the gamer module is not being told
the full information about games won/lost/played and so it can
happen that it tries to write a -1 to the database.  In order to
stop this checks have been added which write watchdog statements
to alert admin!